Fixed several issues
Removed unnecessary hidden fields.

Fixed the Event Type filtering not working for past events and when navigating using pagination.

Removed the Event Type dropdown from the Agenda page as per the sponsors' request.

Fixed an issue with the Agenda page's date and workshop filter not working at the same time.