All features implemented
Implemented the following:

- Update the compose email page to set which categories the email that is being sent is under

- Update the my account page to allow opting out of the several types of emails

- Update the edit user page to allow opting out of several types of email
Update all of the logic that sets space_status (there is logic for this on the my account page, renew membership button, and edit user page). Currently space_status has 4 values: current, expired, current_no_email, and expired_no_email. Revert space_status to just be current or expired and instead use the new columns in the database to store which types of emails are opted out of.